requirements.txt
Django==1.10.5
django-appconf==1.0.2
django-bootstrap3==8.1.0
django-compressor==2.1
django-extra-views==0.8.0
django-nose==1.4.4
django-sass-processor==0.5.3
hashids==1.2.0
libsass==0.12.3
nose==1.3.7
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
